+++ CHANGELOG.md
@@ -166,6 +166,11 @@
   [#1099](https://github.com/Kong/gateway-operator/pull/1099)
 - Remove the owner relationship between `KongService` and `KongRoute`.
   [#1178](https://github.com/Kong/gateway-operator/pull/1178)
+- Check whether an error from calling Konnect API is a validation error by
+  HTTP status code in Konnect entity controller. If the HTTP status code is
+  `400`, we consider the error as a validation error and do not try to requeue
+  the Konnect entity.
+  [#1226](https://github.com/Kong/gateway-operator/pull/1226)
 
 [kubebuilder_3907]: https://github.com/kubernetes-sigs/kubebuilder/discussions/3907